improve robustness to broken input files
One can craft a Bloom filter file in which the number of hash functions
(k) is an unsigned integer larger than the maximum signed integer on the
given platform. This will cause a crash later when trying to make() a
slice with that overly large size, as the length of a slice in Go is
limited by the maximum value of (signed) int.
It should be unlikely to create such a filter using the regular
initializer, but such as filter can be read from a corrupted file.
Here, we fail early at load time when we notice that this value exceeds
the given limit.